==> Scripts/01_data_retrieval.ipynb <==
import os
import numpy as np
import pandas as pd
from collections import defaultdict
%load_ext sql

# Database Connection String

connection_string = "postgresql://Avi@localhost/CHEMBL_34"

# Directory Setup

curated_data_directory = '/Users/Avi/Dissertation/Data/Curated/Raw'
non_curated_data_directory = '/Users/Avi/Dissertation/Data/Non_Curated/Raw'
os.makedirs(curated_data_directory, exist_ok=True)
os.makedirs(non_curated_data_directory, exist_ok=True)

# Curated Data

def gather_data_for_size(minAssaySize=1, maxAssaySize=100000, 
                         onlyDocs=True, removeMutants=True, noDuplicateDocs=True, 
                         onlyHighConfidence=True, readout='IC50'):
    """
    Gather curated data from the CHEMBL database based on specified criteria.
    """
    %config SqlMagic.feedback = False
    %sql $connection_string

    # Create temp_assays table
    %sql drop table if exists temp_assays
    %sql \
    select assay_id, assays.chembl_id assay_chembl_id, description, tid, targets.chembl_id target_chembl_id, \
              count(distinct(molregno)) cnt, pref_name, assays.doc_id doc_id, docs.year doc_date, variant_id \
            into temporary table temp_assays  \
            from activities  \
            join assays using(assay_id)  \
            join docs on (assays.doc_id = docs.doc_id)  \
            join target_dictionary as targets using (tid) \
            where pchembl_value is not null \
            and standard_units in ('nM')  \
            and standard_relation = '=' \
            and target_type = 'SINGLE PROTEIN' \
            group by (assay_id, assays.chembl_id, description, tid, targets.chembl_id, pref_name, \
                      assays.doc_id, docs.year, variant_id) \
            order by cnt desc;

    # Apply filters
    if onlyDocs:
        %sql delete from temp_assays where doc_date is null or doc_date < 2000;

    if removeMutants:
        %sql delete from temp_assays where variant_id is null and (lower(description) like '%mutant%' or lower(description) like '%variant%' or lower(description) like '%mutation%');       

    if onlyHighConfidence:
        %sql alter table temp_assays rename to temp_assays_1;
        %sql select ta1.* into temporary table temp_assays from temp_assays_1 ta1 join assays using(assay_id) where confidence_score = 9;
        %sql drop table temp_assays_1;

    # Create goldilocks table
    %sql drop table if exists goldilocks
    %sql select assay_id, tid, molregno, standard_type into temporary table goldilocks from activities join temp_assays using (assay_id) where pchembl_value is not null and cnt <= :maxAssaySize;

    # Create tmp_counts table
    %sql drop table if exists tmp_counts
    %sql select assay_id, g.tid, count(distinct molregno) cnt, doc_id into temporary table tmp_counts from goldilocks g join assays using (assay_id) group by (assay_id, g.tid, doc_id) order by cnt desc;
    
    %sql drop table if exists goldilocks_counts

    if noDuplicateDocs:
        %sql select distinct on (tid, doc_id) assay_id, tid, doc_id, cnt into temporary table goldilocks_counts from tmp_counts where cnt >= :minAssaySize and cnt <= :maxAssaySize;
    else:
        %sql select distinct on (tid, assay_id) assay_id, tid, doc_id, cnt into temporary table goldilocks_counts from tmp_counts where cnt >= :minAssaySize and cnt <= :maxAssaySize;
        
    %sql drop table if exists goldilocks_target_counts
    %sql select tid, count(distinct assay_id) target_assay_count, sum(cnt) target_compound_count into temporary table goldilocks_target_counts from goldilocks_counts gc group by (tid);

    # Retrieve the raw data
    raw_data = %sql \
      select assay_id, assay_chembl_id, temp_assays.tid, target_chembl_id, standard_type, pchembl_value, \
        md5(textin(record_out((assay_type, assay_organism, assay_category, assay_tax_id, \
                  assay_strain, assay_tissue, assay_cell_type, assay_subcellular_fraction, bao_format, temp_assays.variant_id)))) assay_conditions_hash, \
        molregno, cid.chembl_id compound_chembl_id, canonical_smiles \
        from activities \
        join temp_assays using (assay_id) \
        join goldilocks_target_counts using (tid) \
        join goldilocks_counts using (assay_id) \
        join assays using (assay_id) \
        join chembl_id_lookup cid on (entity_type='COMPOUND' and molregno=entity_id) \
        join compound_structures using (molregno) \
        where target_assay_count >= 20 and target_compound_count >= 1000 \
              and pchembl_value is not null \
              and standard_units = 'nM'  \
              and data_validity_comment is null  \
              and standard_relation = '=' \
              and standard_type = :readout;

    # Process and save data
    min_dataset_size = 10 
    collected_data = defaultdict(list)
    assay_conditions_accum = defaultdict(set)
    assay_accum = defaultdict(set)

    for (_, assay_chembl_id, _, target_chembl_id, standard_type, pchembl_value, 
         assay_conditions_hash, _, compound_chembl_id, canonical_smiles) in raw_data:
        collected_data[(target_chembl_id, assay_conditions_hash)].append((assay_chembl_id, compound_chembl_id, canonical_smiles, pchembl_value))
        assay_conditions_accum[target_chembl_id].add(assay_conditions_hash)
        assay_accum[(target_chembl_id, assay_conditions_hash)].add(assay_chembl_id)

    for target_chembl_id, conds in collected_data.keys():
        rows = collected_data[(target_chembl_id, conds)]
        if len(rows) < min_dataset_size:
            continue
        
        cols = np.array(rows).transpose()
        cond_idx = list(assay_conditions_accum[target_chembl_id]).index(conds) + 1
        
        columns = ['assay_chembl_id', 'compound_chembl_id', 'canonical_smiles', 'pchembl_value']
        df = pd.DataFrame(dict(zip(columns, cols)))
        fname = os.path.join(curated_data_directory, f'target_{target_chembl_id}-{cond_idx}.csv.gz')
        df.to_csv(fname, index=False)

# Non_Curated Data

def gather_data_for_all_data():
    """
    Gather all data from the CHEMBL database without additional filtering.
    """
    %config SqlMagic.feedback = False
    %sql $connection_string

    # Create temp_assays table
    %sql drop table if exists temp_assays
    %sql \
    select assay_id, assays.chembl_id assay_chembl_id, description, tid, targets.chembl_id target_chembl_id, \
              count(distinct(molregno)) cnt, pref_name, assays.doc_id doc_id, docs.year doc_date, variant_id \
            into temporary table temp_assays  \
            from activities  \
            join assays using(assay_id)  \
            join docs on (assays.doc_id = docs.doc_id)  \
            join target_dictionary as targets using (tid) \
            where pchembl_value is not null \
            and target_type = 'SINGLE PROTEIN' \
            group by (assay_id, assays.chembl_id, description, tid, targets.chembl_id, pref_name, \
                      assays.doc_id, docs.year, variant_id) \
            order by cnt desc;

    # Create goldilocks table
    %sql drop table if exists goldilocks
    %sql select assay_id, tid, molregno, standard_type into temporary table goldilocks from activities join temp_assays using (assay_id) where pchembl_value is not null;

    # Create tmp_counts table
    %sql drop table if exists tmp_counts
    %sql select assay_id, g.tid, count(distinct molregno) cnt, doc_id into temporary table tmp_counts from goldilocks g join assays using (assay_id) group by (assay_id, g.tid, doc_id) order by cnt desc;
    
    %sql drop table if exists goldilocks_counts
    %sql select distinct on (tid, doc_id) assay_id, tid, doc_id, cnt into temporary table goldilocks_counts from tmp_counts;
        
    %sql drop table if exists goldilocks_target_counts
    %sql select tid, count(distinct assay_id) target_assay_count, sum(cnt) target_compound_count into temporary table goldilocks_target_counts from goldilocks_counts gc group by (tid);

    # Retrieve all data
    raw_data = %sql \
      select assay_id, assay_chembl_id, temp_assays.tid, target_chembl_id, standard_type, pchembl_value, \
        molregno, cid.chembl_id compound_chembl_id, canonical_smiles \
        from activities \
        join temp_assays using (assay_id) \
        join goldilocks_target_counts using (tid) \
        join goldilocks_counts using (assay_id) \
        join assays using (assay_id) \
        join chembl_id_lookup cid on (entity_type='COMPOUND' and molregno=entity_id) \
        join compound_structures using (molregno) \
        where pchembl_value is not null;

    # Process and save data
    collected_data = defaultdict(list)
    for (_, assay_chembl_id, _, target_chembl_id, _, pchembl_value, 
         _, compound_chembl_id, canonical_smiles) in raw_data:
        collected_data[target_chembl_id].append((assay_chembl_id, compound_chembl_id, canonical_smiles, pchembl_value))

    for target_chembl_id in collected_data.keys():
        rows = collected_data[target_chembl_id]
        cols = np.array(rows).transpose()
        
        columns = ['assay_chembl_id', 'compound_chembl_id', 'canonical_smiles', 'pchembl_value']
        df = pd.DataFrame(dict(zip(columns, cols)))
        fname = os.path.join(non_curated_data_directory, f'target_{target_chembl_id}.csv.gz')
        df.to_csv(fname, index=False)

# Run the Functions

gather_data_for_size()
gather_data_for_all_data()
